feat(Runtime): Expose mirrored skeleton's skeletons, update hip pinning logic, and remove interface support
Summary:
Expose mirror skeleton's original and copied skeletons. This allows us to remove OVRBody that was used to check if the data was valid on the mirrored prefab.

Updated hip pinning logic so that it no longer depends on data from IOVRSkeletonDataProvider, and applies constraints manually instead of letting Unity tick the positional constraint. This is required to get the offset between the hip pinned position and the tracked position to be applied to the rest of the body.

Removed InterfaceSupport as it's no longer used by the HipPinningLogic.

Reviewed By: sohailshafiiWk

Differential Revision: D41390108

fbshipit-source-id: ea07a4fa19c65804649f06e4cc162824591a6186