feat(Runtime): Allow creating duplicate skeleton from animator
Summary: Option to copy original transforms, useful if one wants to create a copy of an animator's starting T-pose transform data before it moves, as opposed to relying on a separate animator stuck in T-pose. This means that retargeting can rely on a single animator for its original T-pose (which it can copy during initialization) and use the same one for retargeting after the fact.

Reviewed By: andkim-meta

Differential Revision: D43557612

fbshipit-source-id: 9f9b10b16800ed8170d473f1cba064a885e23e28