fix(Samples): Old scene names use the word Legacy
Summary:
Old scenes use the suffix "Legacy" while the new ones do not. This means the default scenes should also not use the suffix "V2." As an example, `MovementAura `is the new scene while `MovementAuraLegacy` is the old one.

The `SceneSelectMenuV2` prefab has been updated to use the new scene names.

Reviewed By: andkim-meta

Differential Revision: D45915719

fbshipit-source-id: 83d5e88ae25cf9cb3ec01c48045131fbec5e3df8