feat(Runtime): GroundingConstraint supports Animator; fix IsValid functions
Summary:
GroundingConstraint now supports animator.

Some of the constraints' validity check checked if either the animator or skeleton was null and returned false if one of them is null. That is not correct since a skeleton can be null while the animator is used to retrieve the bones. In fact that kind of check caused the grounding constraint to not work. The constraint is invalid only if *both* fields are null.

Reviewed By: andkim-meta

Differential Revision: D46374485

fbshipit-source-id: 16fb42a679d0e035de6a3d9d7e275a7d28f93944