fix(Runtime): Fix proxy, retargeter update check logic in CheckForSkeletalChanges
Summary:
The code that checks to see if the rig should not be disabled an re-enabled in `CheckForSkeletalChanges` was difficult to read. This diff cleans it up.

If one is using proxy transforms to track the skeleton, and those proxy transforms have not been updated, then the rig should be enabled and disabled. Furthermore, the rig should not go through this process if the retargeting layer has not been updated. This can be checked by negating `HasRetargeterBeenUpdated`.

Reviewed By: andkim-meta

Differential Revision: D46465473

fbshipit-source-id: 7a08cfbfc50aa64d2bdceeb3418b88b20b4f54fa