fix(Runtime): Don't check for animator initialized state in constraints
Summary:
Checking for `isInitialized` on Animator caused the animator to not be used by constraints, since a constraint might initialize before the Animator has a chance to. It should be fine to access an animator's bone transforms before initialization.

I've accessed bone transforms like the Hip on an un-initialized Animator and the Animator still returns a valid reference.

Reviewed By: tseglevskiy

Differential Revision: D46580909

fbshipit-source-id: ea20f616d296adc1c9a24eee47b860056ad7a713