fix(Runtime): Disable constraint game objects until data has been prepared
Summary:
In general, the constraints that obtain references at runtime should be kept off. If they are enabled accidentally on rigs, then their initially null references might be accessed by Unity's animation system, resulting in an error. This diff prevents that.

It also means that these constraints can be added to rigs that belong to RigBuilders that are on by default, which is normal compared to our previous strategy of keeping RigBuilders off until the character skeleton has initialized. It should be up to an individual constraint to handle the edge case where it initializes before its dependent skeleton does.

Since some of our scenes use OVRCustomSkeleton though, which re-arranges the hiearchy of bone transforms, we should still keep the RigBuilder off in those scenes until that process is done. This is the case even if the constraint code is more resilient against the aforementioned initialization issues.

Note that hip pinning's AnimationRigSetup needs to call regenerate data on all of its constraints, so this diff adds those constraint references as seen in the following picture:
{F1024097236}

Also, in the case where the skeleton initializes when `RunInitialSetup` runs, we should initialize our constraints there too.

Reviewed By: andkim-meta

Differential Revision: D46612284

fbshipit-source-id: 9dd0cdf909f083d1c5e6c276a81a427b6f2c1a09