fix(Runtime): OnValidate function check should make sure editor is not playing
Summary: The `OnValidate` error in our constraints was meant to be triggered when the editor was not running, since it should check to make sure the constraint was disabled before the scene starts. This diff makes sure it does not trigger when the editor playing, like when you change the weight of the constraint and have the same warning appear over and over again.

Reviewed By: andkim-meta

Differential Revision: D46660499

fbshipit-source-id: 20aef83dd3c0691a6ef1dcede29a41bd321740f8