fix(Runtime): RetargetingConstraint uses AvatarMask instances
Summary:
`RetargetingConstraint` uses an `AvatarMask` instance to prevent users from modifying the original field. Otherwise if a person calls `SetHumanoidBodyPartActive` on the field itself (through the original getter), they would modify the original mask. An instance is a copy, which means the original mask would not be modified.

Furthermore, an editor script has been added to regenerate the `AvatarMask` instance if the field was modified.

Reviewed By: andkim-meta

Differential Revision: D46815827

fbshipit-source-id: 5ef78f6e02eb863a4435dfc033d6137053c937d8