feat(Runtime): Grounding constraint finds hips and computes offsets at edit time
Summary:
Grounding constraint is now required to find the hips and leg offsets before the scene runs. This means we don't need a set up function to initialize everything.

Also, we make it so that the legs do not need to be re-parented. In general, we should avoid re-parenting since that might negatively character rigs.

This means the mirrored character should disable re-parenting for the leg bones:

 {F1039137567}

Reviewed By: andkim-meta

Differential Revision: D47111671

fbshipit-source-id: 3bbc2a2c710f29cb5d565293bab40343e7d37bf4