fix(Runtime): Fix OVR bone pairs issue, cast to BodyTrackingBoneId when creating visual objects
Summary:
The bone pairs used for  visual objects was creating an axis from the tip the little tip to ring finger, which is wrong. The bone pairs for each axis should exist on the same finger.

Also when OVRSkeleton bone objects are created, their names are based on the index cast to `BoneIdToHumanBodyBone`. This is because the original body tracking enum values overlap with hand values, and if they are used to create game object names then it looks like we have hand tracking bone visuals in the scene. `BoneIdToHumanBodyBone` is more restricted and doesn't have that problem.

Reviewed By: andkim-meta

Differential Revision: D47455368

fbshipit-source-id: 916511f7e0ac000e0f49ccfcd6574051cc95a91e