feat(Runtime): Allow duplicate tuples to be rendered for animator bones
Summary: This allows duplicate tuples (i.e. same first bone, different second bone) to be rendered with the line renderer. This is possible because the custom data variables use a map that maps tuple instance to line renderer. Before it was using a map that was used HumanBodyBones as a map, which means that multiple Tuples with the same starting bone would map to the same line renderer.

Reviewed By: andkim-meta

Differential Revision: D47594159

fbshipit-source-id: fb25f782699725a02b4f230859d0795c9cf2f458