fix(Samples): Fix legs of spawnable green robot
Summary:
The robot's legs were being straightened via Multi-Position and Multi-Rotation constraints. Those constraints were complicated to set up, and did not look correct when applied to robots that were reflected across z.

Two-bone IK is now being applied on the legs of the retargeted robot and the green one that is spawned.

Reviewed By: andkim-meta

Differential Revision: D47978393

fbshipit-source-id: 7a4aa494278877a09b5c9d47329ae8ea1a8d9923