fix(Runtime): Remove LateUpdate correction mask
Summary:
The goal is to avoid using any correction masks. The correct approach is to adjust *all* bones. Assuming the user wants that.

We don't want to remove the avatar component anymore via retargeting, because that would prevent Humanoid animations from running. We were removing the avatar at runtime to avoid human-space position restrictions originally, but with our LateUpdate fix we can get around that.

Reviewed By: andkim-meta

Differential Revision: D47458515

fbshipit-source-id: b7cf927b529316f6849d71df3df7e8325237c9c3