feat(samples): Adds menu to toggle animation on retargeted character
Summary:
Adds a menu to allow playing a custom animation on the retargeted character. This will show people how to modify the retargeting mask to ignore joints that an animation will play on.

The animated character has an `AnimatorController` that will need to activate the animation clip in the first place, and the `CustomAnimToggle` also accesses the controller.

Reviewed By: andkim-meta

Differential Revision: D48755430

fbshipit-source-id: edf9d8c2fc77af5cec133c5a17810440a4452c1d