fix(Samples): Disable _rigToggleOnFocus
Summary:
`_rigToggleOnFocus` was used to prevent certain constraints from running during the de-focus even when the app was paused, one such constraint being the deformation constraint. It operates by disabling the rig, which causes the character to revert to the default "motorcyle pose."

Deformation used to keep deforming the character, however now it checks to make sure that the bone data is valid (which can happen during pausing). This means we don't need to disable the rig anymore since deformation won't keep running when the user pauses.

Reviewed By: andkim-meta

Differential Revision: D49793594

fbshipit-source-id: 5737438c34cf18197dc2946adb3252da53a53bf6