fix(Runtime): Allow controlling blendshape example's constraints via skeletal state
Summary:
The blendshape mapping character has constraints that constantly add an offset to the head and neck bones per frame to make the character look better. During each frame:

1. OVRSkeleton updates the bone positions.
2. Two constraints move the head up and neck down so that the character does not look so compressed. This is done using a per-frame offset.

The problem is that if the user takes the headset off, #1 does not run, but the constraints from #2 will continue to affect the bones. They will continue to add offsets to the head and neck bones for each new frame, causing a strange stretch. Their existence depends on the skeleton updating the bone position *first* so that they can tweak it via an offset.

The constraints should not run if the skeletal data is not valid or not initialized. This will prevent the constraints from running when the user's positions are not being set.

Reviewed By: andkim-meta

Differential Revision: D49867991

fbshipit-source-id: 59562b742253a1a694b270f99433b7e37f9c5969