fix(Runtime): Fix spawning positional problems in retargeting scene
Summary:
The constraint and non-constraint characters need to manage their own spawn offset once they are removed. The spawn offset dictates where the next character spawns, and a removal means that the offset for a specific type of character is decremented.

So if a non-constraint character is removed, the non-constraint offset is affected. This diff adds metadata to each character removed such that when it is removed, its specific offset is affected. Before, `_currentSpawnOffset` (the non-constraint offset) is affected for the removal of both characters, which is wrong and creates a buggy behavior.

Reviewed By: andkim-meta

Differential Revision: D53029902

fbshipit-source-id: 5c8ccac36c6bb6e8e08bdab152a829f6983232cb