refactor (runtime): Locomotion animation fix
Summary:
# Fixed broken jumping animation
Replaces use of `AnimationConstraintBlender` with `StateTransition`.

`AnimationConstraintBlender` is designed specifically to interpolate the weight variable of a rig constraint over a given exit time.
 {F1438263952}

`StateTransition` is designed to interpolate any float property as a state is transitioned into and out over, over a given enter time and exit time.
{F1438239694}

# Renamed rig constraint objects for clarity
This Diff also renames members of a Rig hierarchy to better describe what is going on:
 {F1438267192}

The following GameObjects were renamed for clarity in the "MovementLocomotion" scene, in the "Character" -> "ArmatureSkinningUpdateRetargetLocomotion" -> "Rig" parent object:
- `CaptureAnimationConstraint` -> `RememberAnimatedBodyPose`
- `Deformation` -> `ApplyBodyTrackedBodyPose`
- `PlaybackAnimationConstraint` -> `PartiallyReapplyAnimatedBodyPose`

# Public documentation
This new functionality is important for correctly implementing a Locomotion controller. This change should be documented publicly. The iterating document for that documentation is here:
https://docs.google.com/document/d/1TTeFhorNyGZBirwVUTUEY_rB2wU_J9ZqNrt9D497VJ8/edit?usp=sharing

Reviewed By: andkim-meta

Differential Revision: D53371431

fbshipit-source-id: 5ccb6fdf6ee72dfd9fe816a1f9d4da0455192d42