refactor(Runtime): Remove RetargetingLayer's extra rotation tweaks array
Summary:
`_jointRotationTweaks` was added to `RetargetingLayer` because the base class retargeter's adjustment class did not allow accumulating multiple rotations; it only allowed one rotation value to be used to tweak the retargeted result. Some of our characters require multiple quaternions to be used per joint, which avoids Gimbal lock.

Eventually the base retargeter added a new field to accumulate rotations, so the one in `RetargetingLayer` can now be removed. We recommend using the former in our docs anyway.

The robot needed to have its tweaks moved to the adjustment array from its original location.

From:
 {F1455608571}

To:
 {F1455608701}

Reviewed By: andkim-meta

Differential Revision: D53676626

fbshipit-source-id: 9d0eb29bda9cfea15448636ad1d7f6768111edca