feat(Runtime): Incorporate ExternalBoneTargets
Summary:
Incorporates `ExternalBoneTargets` into one clicks and runtime code. Also fixes robot prefabs to use it.

Furthermore, the spawning code was creating only half body robots. The full body spawning code had a bug which was exposed when that was enabled; it didn't add the lower-body targets. This diff fixes that issue. In addition `GetFallbackJointAdjustments` was adding feet targets, which won't work for upper-body characters (so that was fixed too).

Reviewed By: andkim-meta

Differential Revision: D54145730

fbshipit-source-id: efc3b014ae5b6bafa94d0bbf08ebfc618c417d0c