refactor(Editor): Combine common helper menu code
Summary:
`HelperMenus` and `FullBodyHelperMenus` had a lot of code that was either similar or the same. This diff moves common code into the class that was already be used for that purpose, `HelperMenusCommon`.

Also, `FullBodyDeformationConstraintEditor.cs` was looking up leg weights based on their previous variable name. This causes a null reference exception, so that had to be fixed as well.

Reviewed By: andkim-meta

Differential Revision: D54519721

fbshipit-source-id: 267777650b0dc148ed6f84f48af7d49b537658eb