refactor(Editor): Move common functionality to AddComponentsHelper
Summary:
The functions shared between the runtime and editor code should live in runtime. That's because runtime can depend on the editor assembly, but the editor assembly does not depend on the runtime. This change is a bit of a compromise because some editor code will live in Runtime.

The code here is mostly just a copy and paste into a runtime-version of `HelperMenusCommon` called `AddComponentsHelper`. There were a few tweaks:

1. `AddMainRetargetingComponent` affects `_bodySectionToPosition` for half body characters, before it was doing it for both.
2. `AddBasicAnimationRiggingComponents` is being used by the runtime code, so it checks for the existence of the editor. In the next diff, all of the runtime code will use functions in `AddComponentsHelper`, which will check for the editor.

The editor check will fail during compile time since Editor calls cannot be used during that time. So we have to use #if statements. The next diff will fix that and sort of gets around the compromise of having editor code in Runtime.

Reviewed By: andkim-meta

Differential Revision: D54568704

fbshipit-source-id: 9e38c6d2707a378d4da1528fc54f11374355be57