feat(Runtime): Limit deformation's squash and stretch
Summary:
Deformation uses spine correction to move the spine bones to match the user's head. If the user is too small relative to a very tall character, the character will become squished. If the user is too tall relative to a very short character, the character will become stretched. There is also logic to affect leg bones.

The amount of squashing and stretching can now be dialed back with limits. While this might result in a character that doesn't track a person's head (or other limbs such as the legs) as accurately, we should leave that up to the user. We provide disclaimers in the forms of comments and tooltips to warn them.

Reviewed By: andkim-meta

Differential Revision: D54764499

fbshipit-source-id: 1b5de542c36b367d558f0d4a496cbe85d8a9db3c