feat(Samples): Show hands or controllers with disembodied character
Summary:
In locomotion's "disembodied character" mode, the character mesh moves away from you. Since the character's fingers provide a visual guide to pressing the UI buttons, the lack of a character mesh in your vicinity makes the buttons very difficult to press.

The "Disembodiment" object now toggles both the controller and hands meshes (which active depending on the input mode the user has selected) so in disembodied character mode, the ISDK meshes show the user how far their hands (or controllers) are from buttons even if the character moves away.

 {F1474715031}

Reviewed By: mvaganov-fb

Differential Revision: D55506667

fbshipit-source-id: 0a51647e86ce0b0209f5e921a5ec86da10a96397