fix(Runtime): Fix edge case with AnimatorBoneVisualizer's TryGetBoneTransforms
Summary:
# The problem

`TryGetBoneTransforms` tries to find the `BoneTuple` from an animator. If the second bone in that tuple maps to an invalid value (`HumanBodyBones.LastBone`), it attempts to find the first child of the first bone in the tuple.

In some characters, the first bone of the tuple might not have children, so the original edge case that the code was trying to solve for has its own edge case. Calling `GetChild(0)` will cause an error. This can happen if the bone visualizer attempts to create the pair for `LeftIndexDistal`(first bone)-`LastBone`(second bone; invalid), and `LastIndexDistal` has no child since it's the last bone in that finger.

# The solution
If the second bone tuple is invalid, also make sure the first bone has children. If not, then assigned `null` to both transforms that are supposed to be assigned and return `false`.

Reviewed By: mvaganov-fb

Differential Revision: D55643677

fbshipit-source-id: 9cc739d1de7729c759470c060b9708d324b19b3c