feat(Runtime): Deformation allows straightening spine
Summary:
This diff introduced an advanced feature to straighten the spine based on the original character's spine bone offsets. These offsets are in the local space of the bone being straightened, which makes the math simple. There is also a parameter to restrict the amount of spine bones being affected.

While this removes the hunched-back look of a character, it will not represent the spine accurately with respect to the body tracking. Proper warnings are provided in the form of comments.

Some variable names or comments have also been cleaned up in this diff. They related to retargeting and deformation adjustments, which relate to this diff.

Reviewed By: allenjl-meta

Differential Revision: D55400563

fbshipit-source-id: 9d3aed0c182e57734e39be9ff24a83b46498ac4c