feat(Runtime): Original spine offsets should be proportional
Summary:
While the original spine offsets can straighten the spine, they should respect the current distance from the head to the hips relative to what it used to be.

This is done via a scale that is a the division of the new hips-to-head distance by the original hips-to-head distance. As an example, if the character's spine is 0.5 meters, and the original character's spine is 1.0 meters, the original spine offsets should be scaled by 1.0/0.5 = 1/2.

Reviewed By: andkim-meta

Differential Revision: D55980508

fbshipit-source-id: 1f93e946b3aa9cf2a668d271ccaa669547ff3945