feat(Runtime): Add support for automated custom bone mappings in retargeting
Summary:
BREAKING CHANGE: Updated bone mappings requires V63 SDK.

V63 SDK introduces the OVRHumanBodyBonesMappingsInterface, which can be used to create custom bone pair mappings for retargeting. This feature adds support for that capability, adding the ability to set the bone mappings based on which optional bones are present on the Unity humanoid.

The automatic mapping maps the HumanBodyBones.Chest bone to the OVRSkeleton.SpineMiddle bone, for a better result with the spine retargeting on most characters. However, if there is a missing UpperChest bone, the OVRSkeleton.SpineUpper bone will be mapped instead. These mappings can be modified in the inspector under Retargeted Bone Mappings.

Reviewed By: sohailshafiiWk

Differential Revision: D55996107

fbshipit-source-id: af32288b922282f595bd653852f051b6171738a1