fix(Samples): Fix initial rotation issue with hip pinning
Summary:
The hips initial rotation should be cached, and not computed at realtime. This prevents the issue where the hips rotation at startup is not angled to the left or right during scene initialization.

Also sets hip correction type to head, since that looks better on the hip pinned character. The hips should be controlled by the constraint.

Reviewed By: andkim-meta

Differential Revision: D56578240

fbshipit-source-id: 478c014abc19b81f63b8cf0a22c774b65ab0871d