fix(Runtime): Properly align feet in hip pinning sample
Summary:
# The problem
A previous code change aligns the hips properly in hip pinning, and partially fixed the issue reported in https://github.com/oculus-samples/Unity-Movement/issues/64.

The feet are a little complicated, and can still be mis-aligned with the body's initial foot alignment if a person is not facing forward during scene startup. The feet targets are first aligned with `GroundJob`, then two-bone IK runs to align the body tracking legs (down to the feet) based on those feet targets.

The problem is that two-bone IK records rotational offsets between the feet targets and body tracking feet during its initialization. This means if a person is not facing "forward" (+Z), then there will be an offset between the two with respect the rotation around the Y-axis.

# The solution

`TargetOffsetCorrection.cs` eliminates the this Y-axis rotation offset by aligning the feet with the targets until animation rig (that controls two bone IK) is enabled, in which case the rig can be relied upon to ground and align the feet. Once the rig initializes, two bone IK will effectively not see a rotation offset around the Y-axis.

 {F1501705145}

 {F1501705392}

Reviewed By: andkim-meta

Differential Revision: D56649703

fbshipit-source-id: b953830a288f118227405212486f69aec8d5241b