fix(Runtime): Enforce proper rest pose when using animation rigging retargeting
Summary:
# The Problem
`RetargetingLayer` relies on `OVRUnityHumanoidSkeletonRetargeter`'s `Start` function to store or cache the rest pose of the character so that it can compute the rotation offsets necessary to align it with the body tracking rest pose. This is based on the assumption that `Start` runs before the character's bones move, which is mostly valid because the retargeting constraint that animates the character is disabled until skeletal information is available.

The problem is that if the character's `RigBuilder` is enabled by default, the character might take Unity "muscle pose" before the logic that stores its rest pose can run, which is a race condition breaks the assumption mentioned. Animation rigging will force the character into muscle pose (like it's sitting on a motorcycle) even if no jobs or animations are running, so the cache rest pose is effectively the muscle pose. Using the muscle pose compared with the body tracking rest pose might result in odd rotation offsets for many joints, especially the fingers.

Furthermore, some characters seem to have rotation value set to transforms *above* the hips too when animation rigging turns on, which causes a related problem. When computing the offset for the character's rest pose after the first frame, it must be aligned with its original cached rest pose. If there is a rotation above the hips that is say 90 degrees when animation rigging turns on, the character will look like it is resting on the ground. I've noticed this with one character, which suddenly has a "root" transform with a 90 degree rotation that animation rigging enforces by default. One can get around this by creating a prefab that strips this rotation and uses that as the retargeted character, but the code should handle this edge case by storing the original rotation values from the hips upwards when the character is at rest, then re-applying this to force the character into a rest pose.

# The solution
All characters, whether spawned or created through our one-click tool, must have their `RigBuilder` disabled by default so that the rest pose can be captured by our retargeting system without animation rigging affecting it beforehand.

Furthermore, if the character has any transformations of *bones* above the hips that have a non-identity rotation value, those must be reset too to their original values from `Start`. We should not affect the transformation on the topmost transform containing the animator because a person might want to move the character around or rotated it. This will force the skeleton of the character to match the upright posture of the cached rest pose, because they both must be aligned under the parent transform containing both.

The solution that resets rotations above the hips *also* needs to detect if an offset is about to happen. Unfortunately, the base class's variables that tell us this are private and not visible to the inheriting class, which means the inheriting class must create duplicate variables that tell when offsets are about to be recreated per frame. These variables are based on detected scale changes or if the underlying `OVRSkeleton` change count has incremented, both of which tell the underlying retargeter to recompute offsets.

All of these changes were done to make our retargeter work nicely with animation rigging.

Reviewed By: andkim-meta

Differential Revision: D56893989

fbshipit-source-id: 6850ad568c010bde0d972f4c9eaedaa050a26f4f