fix(Runtime): Allow fixing spine rotations in deformation's ApplyOriginalSpineOffsets
Summary:
The parameter to control the influence of the original spine positions does not affect the rotation of the bones in the spine. Imagine if there is an axis between the spine bone and the lower chest bone that sits above it. If the lower chest bone moves, this axis should be rotated.

This is what `RotateJointBasedOnNewPosition`, and the entirety of this logic is based on `FixRotationsOfJoints` as seen in the FABRIK code in `AnimationUtilities`.

Once a joint is rotated, we must be careful that its children's world positions and rotations are restored to what `ApplyOriginalSpineOffsets` indicates they should be (cached transform values). So before the rotation is fixed for spine joint `x`, the transform values for its children are restored to the cache values.

The only spine bone which is left uncorrected is the hip bone. That bone sits between the lower and upper body. If its rotation is adjusted based on the new lower spine position, the legs bones might be rotated out of position.

Reviewed By: andkim-meta

Differential Revision: D57361206

fbshipit-source-id: d548feaf4819e70ccb72daf5dbe1c9edfc2c3ffc